requirements.txt
openai<=0.28.1
langchain
replicate
pillow
SpeechRecognition
geocoder
overpy